refactor: make ai-recommendation as a file
