+++ src/app/recommendation/page.tsx
@@ -12,7 +12,6 @@ import useFetch from '@/hooks/use-fetch'
 import { api } from '@/utils/api'
 import { getMapId } from '@/services/map-id'
 import {
-  confusingInputChat,
   initialRecommendChat,
   lastChat,
   noInfoLocationChat,
@@ -23,6 +22,7 @@ import useUserGeoLocation from '@/hooks/use-user-geo-location'
 import GptIntroModal from '@/components/kakao-map/gpt-intro-modal'
 import VisuallyHidden from '@/components/common/visually-hidden'
 import Icon from '@/components/common/icon'
+import { handleAIRecommendation } from './ai-recommendation'
 
 const Recommendation = () => {
   const { data: user } = useFetch(api.users.me.get, {
@@ -32,6 +32,7 @@ const Recommendation = () => {
     data: recommendationUsage,
     status,
     revalidate,
+    refetch,
   } = useFetch(api.gpt.usage.get, {
     key: ['recommendation-usage'],
   })
@@ -68,132 +69,22 @@ const Recommendation = () => {
     return { location: true, allow: allowLocation }
   }
 
+  const refetchUsage = () => {
+    revalidate(['recommendation-usage'])
+    refetch()
+  }
+
   const askToAI = async (suggestion?: string) => {
-    if (availableCount <= 0) {
-      setChats((prev) => [
-        ...prev,
-        { ...usageCapReachedChat, suggestionKeywords: [] },
-      ])
-      return
-    }
-
-    setIsFetching(true)
-
-    const isLastChat = (index: number) => chats.length + 1 === index
-
-    setChats((prev) => [
-      ...prev,
-      {
-        type: 'gpt-stream',
-        value: '',
-        suggestionKeywords: [],
-      },
-    ])
-
-    try {
-      const x = String(userLocation?.longitude || '')
-      const y = String(userLocation?.latitude || '')
-      const recommendationApi =
-        process.env.NODE_ENV === 'production'
-          ? api.gpt.restaurants.recommend
-          : api.gpt.restaurants.recommend.test
-      const response = await recommendationApi.get(suggestion || input, x, y)
-
-      const reader = response.body.getReader()
-      const decoder = new TextDecoder()
-      const loopRunner = true
-      while (loopRunner) {
-        const { value, done } = await reader.read()
-        if (done) {
-          break
-        }
-
-        const decodedChunk = decoder.decode(value, { stream: true })
-        setIsFetching(false)
-
-        try {
-          const isError = /event: error/g.exec(decodedChunk)
-          if (isError) {
-            setChats((prev) =>
-              prev.map((chat, index) => {
-                if (isLastChat(index)) {
-                  return confusingInputChat
-                }
-                return chat
-              }),
-            )
-            return
-          }
-
-          const parseTextData = (text: string) => {
-            const arr = []
-            const regex = /data: ({.*?})\n/g
-            const matches = text.match(regex) || []
-            for (const match of matches) {
-              const jsonData = JSON.parse(
-                match?.replace('\n', '').replace('data: ', ''),
-              )
-              arr.push(jsonData)
-            }
-            return arr
-          }
-
-          const jsonArr = parseTextData(decodedChunk)
-
-          for (const { data, type } of jsonArr) {
-            if (type === 'text') {
-              const character = data
-              setChats((prev) =>
-                prev.map((chat, index) => {
-                  if (isLastChat(index)) {
-                    return {
-                      ...chat,
-                      value: chat.value + character,
-                    }
-                  }
-                  return chat
-                }),
-              )
-            } else if (type === 'json') {
-              setChats((prev) =>
-                prev.map((chat, index) => {
-                  if (isLastChat(index) && chat.type === 'gpt-stream') {
-                    return {
-                      ...chat,
-                      suggestionPlaces: [
-                        ...(chat.suggestionPlaces || []),
-                        data,
-                      ],
-                    }
-                  }
-                  return chat
-                }),
-              )
-            }
-          }
-        } catch (err) {
-          console.log(err)
-        }
-      }
-
-      setChats((prev) =>
-        prev.map((chat, index) => {
-          if (isLastChat(index)) {
-            return {
-              ...chat,
-              suggestionKeywords: ['처음으로'],
-            }
-          }
-          return chat
-        }),
-      )
-
-      setIsFinish(true)
-      revalidate(['recommendation-usage'])
-    } catch (error) {
-    } finally {
-      setIsFetching(false)
-    }
+    handleAIRecommendation({
+      userLocation,
+      question: suggestion || input,
+      availableCount,
+      chats,
+      setChats,
+      setIsFetching,
+      setIsFinish,
+      refetchUsage,
+    })
   }
 
   const sendChat = async () => {
